【后端-注册登录】测试完成
注册测试完成

+++ IEXSYS/src/main/java/org/exp/iexsys/mapper/UserMapper.java
@@ -7,7 +7,7 @@ import org.exp.iexsys.domain.User;
 @Mapper
 public interface UserMapper {
 
-    User selectById(@Param("id") Integer id);
+    User selectById(@Param("id") Long id);
 
     User selectByUsername(@Param("username") String username);
